+++ mycodo/mycodo_flask/utils/utils_math.py
@@ -13,6 +13,7 @@ from mycodo.databases.models import DeviceMeasurements
 from mycodo.databases.models import DisplayOrder
 from mycodo.databases.models import Input
 from mycodo.databases.models import Math
+from mycodo.databases.models import Output
 from mycodo.mycodo_flask.extensions import db
 from mycodo.mycodo_flask.utils.utils_general import add_display_order
 from mycodo.mycodo_flask.utils.utils_general import controller_activate_deactivate
@@ -144,9 +145,10 @@ def math_mod(form_mod_math, form_mod_type=None):
 
         # Collect inputs and measurement name and units
         if mod_math.math_type in ['average',
-                                  'redundancy',
                                   'difference',
+                                  'redundancy',
                                   'statistics',
+                                  'sum',
                                   'verification']:
             if len(form_mod_math.inputs.data) < 2:
                 error.append("At least two Inputs must be selected")
@@ -162,14 +164,48 @@ def math_mod(form_mod_math, form_mod_type=None):
             # Change measurement information
             if form_mod_type.average_input.data and ',' in form_mod_type.average_input.data:
                 measurement_id = form_mod_type.average_input.data.split(',')[1]
-                selected_measurement = get_measurement(measurement_id)
-                if selected_measurement:
-                    conversion = Conversion.query.filter(
-                        Conversion.unique_id == selected_measurement.conversion_id).first()
+
+                if measurement_id == 'output':
+                    output = Output.query.filter(
+                        Output.unique_id == form_mod_type.sum_input.data.split(',')[0]).first()
+                    unit = output.unit
+                    measurement = output.measurement
                 else:
-                    conversion = None
-                _, unit, measurement = return_measurement_info(
-                    selected_measurement, conversion)
+                    selected_measurement = get_measurement(measurement_id)
+                    if selected_measurement:
+                        conversion = Conversion.query.filter(
+                            Conversion.unique_id == selected_measurement.conversion_id).first()
+                    else:
+                        conversion = None
+                    _, unit, measurement = return_measurement_info(
+                        selected_measurement, conversion)
+
+                mod_measurement = DeviceMeasurements.query.filter(
+                    DeviceMeasurements.device_id == form_mod_math.math_id.data).first()
+                mod_measurement.measurement = measurement
+                mod_measurement.unit = unit
+
+        if mod_math.math_type == 'sum_single':
+            mod_math.inputs = form_mod_type.sum_input.data
+
+            # Change measurement information
+            if form_mod_type.sum_input.data and ',' in form_mod_type.sum_input.data:
+                measurement_id = form_mod_type.sum_input.data.split(',')[1]
+
+                if measurement_id == 'output':
+                    output = Output.query.filter(
+                        Output.unique_id == form_mod_type.sum_input.data.split(',')[0]).first()
+                    unit = output.unit
+                    measurement = output.measurement
+                else:
+                    selected_measurement = get_measurement(measurement_id)
+                    if selected_measurement:
+                        conversion = Conversion.query.filter(
+                            Conversion.unique_id == selected_measurement.conversion_id).first()
+                    else:
+                        conversion = None
+                    _, unit, measurement = return_measurement_info(
+                        selected_measurement, conversion)
 
                 mod_measurement = DeviceMeasurements.query.filter(
                     DeviceMeasurements.device_id == form_mod_math.math_id.data).first()